python-cache-demo/infra-service/requirements.txt
requests==2.25.1
requests-cache==1.2.0
Flask==3.0.2
mysql.connector==2.2.9